Allow different I2C slave addresses for LCD
The batch of LCDs we have have one of two different I2C slave
addresses (0x4e or 0x7e).  There seems to be no way of
programatically discovering that address or changing it - at
least no documented way.  The solution here is to create two LCD
objects, one with address 0x4e and one with 0x7e.  All commands
will be sent to both of theses objects.  One commenad will go to
the actual attached LCD and the other will be ignored.
Admittedly not an elegant solution, but it works.

Signed-off-by: Kyle Stein <Kyle.Stein@arm.com>