[m2mclient] Remove unused objects and update API
This patch removes some of the unused resource objects defined in
the m2mclient class. The create_resources function has been replaced
with a add_resource function which will add any necessary objects to
the cloud client.
In addition, the mbed_client variable has been moved to main and will
be passed by ref in all functions and threads that need it.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>