[RED-165] Display abstraction
The display has been abstracted out into a DisplayMan object. The
idea is that the display will be called with methods such as
set_network_fail() or set_sensor_status(...) and the display will
be responsible for translating its state into what LEDs to light,
what to write on the LCD, etc.

Signed-off-by: Kyle Stein <Kyle.Stein@arm.com>
Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>