set power LED to ON before doing anything else
previously, the power LED was set to ON only after the basic
platform init sequence completed.  however, some folks have
encountered init sequence failures.  to help us figure out if
the app's init sequence is failing or if the failure occurs
before the app has control, we need to display the power LED
immediately on startup.

Signed-off-by: Nic Costa <nic.costa@arm.com>