Disable sending sync messages - not used by our fota project
mbedhtrun sends messages by default to the device. However, the
project we are working on here does not accept them so the result
is some error messages. Here I just disable the sending of those
ignored messages.

Also putting the date at the top of the log file is useful.