Added a GIT sha to the serial console.
When main.cpp is recompiled, the a DEVTAG of the SHA and branch are
compiled into the code.  Upon start up, the code version is displayed
in the serial console.  If the compilation happens from within a dev
environment, the DEVTAG conveys reports whose environment was used
to create it.

Signed-off-by: Earl Manning <earl.manning@arm.com>