Generate mbed_app.json from global and local config
Now mbed_app.json is a generated file, not checked in. An optional
local file config_local.json is also not checked in.

The files config_global.json and config_local.json are used to
generate mbed_app.json. The settings in the local config take
precedence.

If a local config does not exist, the global settings are copied,
unchanged, to mbed_app.json.