[boot] Add mBed bootloader to build
This patch adds the mBed bootloader to the build environment. It
removes the prebuilt bootloader from the tools folder and instead
copies in the newly built version.
The .mbedignore file ensures that the mbed-bootloader-* directory
is ignored from the fota-demo build (this otherwise causes duplicate
symbols during link time) and the Makefile invokes the mbed compile
command for each build.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>