[Commander] call on_ready cb for every keystroke
this patch fixes an issue where keystrokes are not echoed back
to the console or processed until the user hits the enter key.
further, only a single keystroke is processed per 'enter'.

For example, if the user types 'help' on the keyboard, the user
must follow this with 4+1 'enter' keystrokes before the command
is fully echoed back to console (4 'enter's) and then finally
processed (1 additional 'enter').

by calling the on_ready handler for each keystroke, the containing
object is prompted immediately.

the reason the callback is executed rather than simply handling
the echo within commander is to allow the containing object
control over the appropriate execution context in which to
handle the event.

Signed-off-by: Nic Costa <nic.costa@arm.com>