[boot] Change bootloader lib location
This patch changes the location of where the bootloader is pulled
from on Github. Changes have been made to the bootloader to support
display of a firmware upgrade on the LCD for the FOTA demo.
The LED indicator also blinks when the bootloader reads, writes,
and verifies the new firmware.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>