[Makefile] fix discovery of toolchain/target/profile
the behavior of 'mbed detect' differs whether the mbed-os
directory has been cloned which affects the rule that
checks if we're building for a valid target.

fix an issue where a 'make distclean' followed by a
'make bootloader' would incorrectly create symlinks to
non-existent files and folders.

fix an issue where the build profile does not exist when
the Make variable is expanded.

Signed-off-by: Nic Costa <nic.costa@arm.com>