increase speed of firmware download
a configurable buffer size can be used to balance between RAM usage
and download rate of new firmware images.

this patch sets the buffer to 8192 from the original value of 2048,
which decreases the time to download a 340k firmware image by
around 45 seconds.  the original time to download was about 1:45
(one minute forty-five seconds) and the new time is around 1 minute.

Very unscientific testing results for additional buffer sizes:
        buffer size: time to download
        2048: 1:46
        4096: 1:18
        8192: 1:00
        16384: :50
        32768: :45
        65536: :44

Signed-off-by: Nic Costa <nic.costa@arm.com>