[main] Adding network scan
RED-263 #comment Add capability to scan for nearby network devices

This patch adds support for scanning nearby devices depending on the
network interface. In the case of WiFi it will scan for nearby APs
and report the number of APs found.
For Ethernet interface the implementation has been left undefined
and could possibly be used in the future to scan for devices on the
local net.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>