[main] Add mBed Cloud support for network scan
This patch adds support for sending data to the mBed Cloud for the
network scan.
The LwM2M object ID is 26242 and we only register a single instance
of this resource. It is NOT periodically updated and only sent once
upon initial boot and scan of the wireless network.

This patch also adds the rapidJSON library to the FOTA demo repo so
that we can format the WiFi AP data into a JSON array which can be
uploaded to mBed Cloud.
The format of the JSON array is a list of objects with each object
containing two key-value pairs:
- the first key is "macAddress" which is the APs IEEE802 MAC address
  formatted as six two-digit octets spearated by colons
  "XX:XX:XX:XX:XX:XX"
- the second key is "signalStrength" which represents the RSSI value
  in decibles as an integer

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>