remove Geo timestamp m2m resource
we don't seem to have a reliable way of initializing the system
clock which means that time(NULL) starts at 0 on system startup.

as a note, after the mbed client registers, the mbed cloud writes
the current time to the resource at /3/0/13 which could possibly
be used to initialize the real time clock via set_time().

Signed-off-by: Nic Costa <nic.costa@arm.com>