[net] Change color of LED during network scan
This patch changes the color of the LED during a network scan once
the network connection has been established.
This is to distinguish between the network being actively connected
and when it's performing some other function before connecting to
the cloud.
The network_status calls in DisplayMan have also been provided to
provide default display messages when invoked. This removes a
number of method calls in main.cpp.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>