Update LEDs on WiFi reconnect
Previously, the cloud LED would stay in the error state (red)
from the time A WiFi reconnect started until the cloud had
successfully reconnected.  This was confusing as the actual
state of the cloud connection was more dynamic in this period.
With this change, the LED will go blank (unregistered) while
WiFi is reconnecting, and then will transition to blinking
yellow until a successful connection is made.

Signed-off-by: Kyle Stein <Kyle.Stein@arm.com>