special case '-' (dash) to delete user lat/long
the world map prioritizes 'user' over 'auto' so that 'auto' is never
displayed if 'user' is configured.

this is fine until the admin wants to remove the 'user' config to
allow the 'auto' location to be displayed.

there is no visible control in the mbed cloud portal to delete a
resource and attempts to delete the resource through the REST api
don't seem to be forwarded to the device, nor is the resource deleted
from the mbed cloud portal.

therefore, in the absence of a functional DELETE verb, we need a
special character to represent delete.  the '-' (dash) was chosen
because this is the value displayed in the mbed cloud portal
for uninitialized resources.

Signed-off-by: Nic Costa <nic.costa@arm.com>