Update manifest-tool
The build fails with the following error:

Traceback (most recent call last):
  File "test.py", line 10, in <module>
    namedtype.OptionalNamedType('blerg', univ.Any())
  File "/home/ryanow01/sandboxes/pyasn1_problem/.tox/pyasn1033/local/lib/python2.7/site-packages/pyasn1/type/namedtype.py", line 117, in __init__
    self.__nonUniqueTagMap = self.__computeTagMaps(unique=False)
  File "/home/ryanow01/sandboxes/pyasn1_problem/.tox/pyasn1033/local/lib/python2.7/site-packages/pyasn1/type/namedtype.py", line 441, in __computeTagMaps
    raise error.PyAsn1Error('Duplicate default ASN.1 type at %s' % (self,))
pyasn1.error.PyAsn1Error: Duplicate default ASN.1 type at NamedTypes(NamedType('baz', Any()), OptionalNamedType('blerg', Any()))

This is because pyasn1 was updated and introduced a bug such that
manifest-tool 1.2rc2 doesn't even start any longer.  That was fixed in
manifest-tool 1.2rc4 so let's update to that.  This was not caught
during the daily build test because we're not starting with an empty
python virtualenv.  We probably should be.