work around mbed cloud client registration failure
the mbed cloud client has a bug that prevents client registration
when the registration message size is greater than 1024.  to
work around the issue, the mbed cloud client team has suggested
to change the coap blockwise payload size from 1024 to 512.
the suggestion seems to have worked.

it is proposed that the recent addition of the auto geo-location
M2M resources pushed the total registration message size over 1024
bytes.

this value is changed in mbed_app.json for mbed-cloud-client
version 1.2.1.

Signed-off-by: Nic Costa <nic.costa@arm.com>