LED map for Odin EVK board
The LED driver behaves significantly differently on the Odin EVK
W2 board than on the K64F.  First, the driver seems to consume
64 bits of control data for each LED instead of the expected 32
bits.  Second, the mapping of the 24 color bits into these 64 bits
does not seem to have a discernable pattern.  This change maps the
bits for the 7 composite colors used in the application indiviually.
The map was derived empirically.  Since the Odin EVK platform is
a short-lived platform for this application, we won't spend time
investigating the driver's behavior, but instead use this emperical
mapping.

Signed-off-by: Kyle Stein <Kyle.Stein@arm.com>