workaround client deregistration/registration cycle
the mbed client doesn't send registration updates frequently
enough to keep the client registered when using a lifetime
of 60 seconds.  updates are sent at 3/4 the configured client
lifetime, but this doesn't seem to be enough time for the
registration update to be processed and acknowledged before
automatic deregistration by the server.

to workaround the problem, we manually send registration
updates via a keep_alive() API in the mbed cloud client on
an interval that is 1/4 of the configured lifetime.

Signed-off-by: Nic Costa <nic.costa@arm.com>