[patches] Apply patches based on target
This commit changes the functionality of how patches are applied
to subdirectories. Under the patches folder exists a COMMON directory
which contains a set of folders that match the libraries checked
out by mbed. Each of these patches is applied for all targets before
any target specific patches are applied.
In addition to the COMMON directory are target specific patch folders
with the same structure. These are applied after the COMMON patches
so that if a specific target needs to make further changes to the
same file, or reverse the changes, it can.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>