[Makefile] Update patch apply
This commit addresses an issue where performing a second mbed update
in the mbed-bootloader directory causes the .patches rule to be
performed again even though the .patches file still exists.
This can be seen by removing the mbed update command on line 150
within the Build/Compile bootloader call; this fixes the issue but
because mbed-bootloader patch updates libraries those will not be
synced.
Instead of relying on the rule being updated we just check if the
file already exists.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>