Another attempt at local mbed_app.json config
the purpose of the change is to allow a developer to maintain
local mbed_app.json overrides, such as wifi-ssid, in such a way
that it will not be lost during the next git-pull.

to address the concern, the 'prepare' target checks for existence
of a file mbed_app_local.json and if present merges the file with
mbed_app.json.

the developer can then maintain mbed_app_local.json with keys
that will either add or overwrite in mbed_app.json.

the resulting merged file is written back to mbed_app.json
so that the normal mbed-cli configuration system behavior is
preserved.

Here is an example mbed_app_local.json that modifies the wifi
ssid and security settings:

$ cat mbed_app_local.json
{
    "config": {
        "wifi-ssid": {
            "value": "\"Guest-AccessNG\""
        },
        "wifi-security": {
            "value": "\"OPEN\""
        }
    }
}

Signed-off-by: Nic Costa <nic.costa@arm.com>