Instantiate Odin before M2M
When the OdinWiFiInterface object is instantiated after the
M2MClient, the EventQueue does not operate as expected.  The
(unconfirmed) theory is that the OdinWiFiInterface constructor
performs some kind of reset on the EventQueue that negates setup
by the M2MClient.  The observed behavior is that M2M timers -
including the keep-alive timer - fail to fire when the Odin
constructor is called after the M2M constructor.  In the example
cloud client application (which uses the easy-connect library) the
OdinWiFiInterface is declared as a global variable, and thus its
constructor is called at initialization (before M2M).

Signed-off-by: Kyle Stein <Kyle.Stein@arm.com>