[main] Adding sensor verbose command
This patch adds a command to enable sensor verbosity when values
are sent to the cloud. For the light sensor value this should always
print, but the DHT sensor may print and error message if for
instance the DHT sensor is not connected.
The command can be invoked by typing:

    verbose <type> [off|on]

The only type supported is currently 'sensors'. Not supplying the
type results in a usage error (for now, see comments in code).
If no off|on parameter is supplied then the current status of the
verbose flag is printed, otherwise the status of the flag is
updated.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>