use git-am to apply patches instead of patch
mbed-update has an extremely unfortunate habit of deleting untracked
files.

this is clearly shown by adding a new .lib to the environment,
say spif-driver.lib.  if the user runs mbed-update without first
committing the new .lib to the git history, then mbed-update deletes
the file instead of doing the sane thing and cloning the repo.

this is obviously not good for a dev workflow, but it also
breaks our build system when applying patches to the bootloader
because we need to add libs to the bootloader.

Signed-off-by: Nic Costa <nic.costa@arm.com>