fix patching of .lib files
work around the following mbed-cli behavior:

* mbed update has an unwanted side-effect of modifying local .lib
files.  since we now use git-am to apply patches, if we apply a
patch to a subdir tracked by a .lib (mbed-bootloader for example),
then any subsequent mbed-update will modify the .lib file
(mbed-bootloader.lib for example) to point to the new SHA created
by running git-am.

this is problematic for a couple of reasons,
1. it makes it easy to carelessly commit an invalid change to a .lib
2. it breaks local builds after a distclean because the SHA in
the .lib no longer exists because the commit-SHA created by git-am
was just local and temporary.

* mbed deploy doesn't notice all changes to .lib files.  if
the library is already downloaded and a change is made to the URL
in a .lib file, mbed deploy does nothing.  mbed deploy seems to
notice changes in the SHA, but not to the repo URL.

Signed-off-by: Nic Costa <nic.costa@arm.com>