[ledman] Place LED logic into class/template
This patch places the logic inside the C calls for ledman into a
Base class which is then derived for each Target (K64F and ODIN).
This allows us to split the generic logic and internal state of
the LED colors and flags into a reusable class and to let the
Target specific classes define how that is translated to hardware.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>