fix K64F failure to reboot during FOTA
added new flag bitfields to keep track of fota progress so
that authorization of fota-download and/or fota-install doesn't
have to occur within the auth callback.

this allows us to call close() on the mbed client before authorizing
fw install in order to disable its reconnect logic which is activated
when we call network->disconnect() due to the ESP8266 ATcmd sync
errors that can occur when the K64 is suddenly rebooted.

Signed-off-by: Nic Costa <nic.costa@arm.com>