Makefile: add certsave target to store cert files
developers frequently run 'make distclean' between builds to
clean up build artifacts and wipe away local code changes in
dependent libraries.  the problem is, distclean also deletes
the certificates used to sign and verify firmware manifests used
during FOTA, leading the developer to need to re-flash the
board then rebuild and FOTA again, taking care not to
distclean in between.

this patch adds a new make target, 'make certsave', which
saves all certificate files into the user's home directory
under a subdir specific to the build target:

        $HOME/.mbedcerts/$TARGET

if the certificate files ever need to be re-generated, for
example after a distclean, then the Makefile first looks
for files contained in this directory and copies them into
the local build directory.  if the files do not exist,
then 'manifest-tool init' is executed as usual.

Signed-off-by: Nic Costa <nic.costa@arm.com>