Update Saving message
On Odin the process of saving the downloaded firmware image takes
roughly ten seconds.  Unfortunately, the operation which takes place
within the cloud client library does not have any context switches
thta would allow us to update the LCD with a progress bar.  This
commit simply changes the LCD message "Saving..." to
"Saving (10s)..." to notify the observer of the expected delay.

Signed-off-by: Kyle Stein <Kyle.Stein@arm.com>