Fix flash driver page size issues
Different versions of the FlashIAP driver return incorrect small
values for the page size.  For example the current driver returns
a value of one byte for the Odin platform.  The previous solution
for this issue was to downrev the FlashIAP driver.  The downreved
driver however returned the wrong pagesize (8 bytes) on the K64F
platform.

This commit removes the downrev of the driver.  It also ignores the
returned value of the flash driver if it indicates a page size of
less than 1024 bytes.  A minimum page size of 1024 is used in
calculations involving page size.

Signed-off-by: Kyle Stein <kyle.stein@arm.com>