bootloader: retry fw image write after failure
the bootloader can detect whether the FOTA image was successfully
copied from SPI flash to internal flash by checking SHA256 of
the stored image.  if this check fails, the bootloader attempts
to write the image again, up to a maximum of MAX_COPY_RETRIES.

this patch increases MAX_COPY_RETRIES from 1 to 3 to try to
work around intermittent write failures.

Signed-off-by: Nic Costa <nic.costa@arm.com>