Downrev flash driver in bootloader
Current flash driver had some issues programming flash where a
write to flash reported a good return code, but the data written
did not match the source.  This seemed to manifest in the write
to the first section of the second bank of flash, but no root
cause was determined for the failure.  Downrevving the driver
fixes the issue, but it should be noted that it is unclear as to
what code in the new driver is causing the issue.

Signed-off-by: Kyle Stein <kyle.stein@arm.com>