Sensor Tuning
The light, temperature, and humidity sensors have now been adjusted to compensate for the case.

The light sensor is tuned to give similar results to a a Dr. Meter LX1330B Digital Lux Meter.

The temperature, and humidity is tuned to give similar results to a Honeywell HHM10B Temp/Humidity Monitor.

Signed-off-by: J. Michael Welsh <mike.welsh@arm.com>