Commander: Use RawSerial instead of Serial
The Serial class uses Stream in order to read from the serial
device, but Stream doesn't allow putc/getc in an interrupt
context which results in the following console message when
running under the debug profile:

        Mutex 0x200010ac error -6: Not allowed in ISR context

RawSerial doesn't have this limitation.

An additional change was required in order to fully fix the
Mutex error, which is that the Commander's internal buffer
that holds characters for processing cannot allocate memory
within the interrupt context.  To fix this, the buffer is
allocated with a reserved capacity and it not allowed to
grow beyond that capacity in the interrupt handler.

Signed-off-by: Nic Costa <nic.costa@arm.com>