Allow LED indicator colors to be configured
The indicator colors can now be set in mbed_app.json.  For example,
adding
        "color-on": {
            "help": "24-bit LED color (BBGGRR) to indicate power on",
            "value": "0xFF00FF"
        },
to the "config" section of mbed_app.json will cause the power on LED
to be bright pink(0xff00ff) intead of green.

color-on, color-in-progress, color-failed, color-success,
color-success-dim and color-sensor-data can all be set.

If no config setting is set, the default color is preserved.

Signed-off-by: Kyle Stein <kyle.stein@arm.com>