fix and re-enable factory reset
the main fix was to set the DigitalIn::mode to PullUp before reading
the button.  the change was required due to a missing pull up resistor
R34 which wasn't populated on the WEM.

a timeout was also added such that the user has to hold the button
down for a configurable number of (default 5) seconds before the
factory reset is carried out.

the timeout can be configured at compile time by setting a value for
config.factory-reset-button-press-secs in mbed_app.json or by
changing MBED_CONF_APP_FACTORY_RESET_BUTTON_PRESS_SECS in main.cpp.

Signed-off-by: Nic Costa <nic.costa@arm.com>